assessment/pyproject.toml
[tool.poetry]
name = "assessment"
version = "0.1.0"
description = ""
authors = ["ArthurMykhailiuk <artur.myhajlyuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
faker = "18.9.0"
psycopg2 = "^2.9.10"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
